grrr windows

+++ test/src_test/include/test_bicop_select.hpp
@@ -43,7 +43,7 @@ TEST(bicop_select, allows_all_selcrits)
 TEST(bicop_select, allow_rotations_works)
 {
   Bicop cop(BicopFamily::clayton, 180, Eigen::VectorXd::Constant(1, 5));
-  auto u = cop.simulate(5e3);
+  auto u = cop.simulate(static_cast<size_t>(5e3));
   FitControlsBicop controls;
   controls.set_family_set({ BicopFamily::clayton });
   cop.select(u, controls);
